mv(env): move update_pwd_by_internal() in declare_arg.c, etc
