+++ Makefile
@@ -41,8 +41,7 @@ SRCS		+=	$(ENVIRONMENT)/clear.c \
 				$(ENVIRONMENT)/separate_key_value.c \
 				$(ENVIRONMENT)/set.c \
 				$(ENVIRONMENT)/set_dup_key_value.c \
-				$(ENVIRONMENT)/unset.c \
-				$(ENVIRONMENT)/update_pwd_by_internal.c
+				$(ENVIRONMENT)/unset.c
 
 DEBUG_DIR	:=	debug
 SRCS		+=	$(DEBUG_DIR)/put.c
