+++ includes/minishell.h
@@ -107,12 +107,8 @@ t_result	env_separate_key_value(const char *const arg, \
 									char **key, \
 									char **value, \
 									t_env_op *op);
-bool		is_key_pwd_oldpwd(const char *key);
 void		env_set(t_env *env, char *key, char *value, t_env_op op);
 void		env_unset(t_env *env, const char *key);
-char		*update_pwd_by_internal(const char *key, \
-									char *value, \
-									t_context *context);
 
 /* destroy */
 void		destroy(t_context context);
